moved some attributes up to Table class from subclasses
