+++ pylytics/library/table.py
@@ -8,7 +8,7 @@ from exceptions import classify_error, BrokenPipeError
 from settings import settings
 from template import TemplateConstructor
 from utils import (_camel_to_snake, _camel_to_title_case, dump, escaped,
-                   classproperty)
+                   classproperty, raw_sql)
 from warehouse import Warehouse
 
 
@@ -127,6 +127,17 @@ class Table(object):
 
     INSERT = "INSERT IGNORE"
 
+    id = PrimaryKey()
+    hash_key = HashKey()
+    created = CreatedTimestamp()
+
+    def __init__(self, *args, **kwargs):
+        values = ', '.join(
+            ["IFNULL(%s,'NULL')" % escaped(c.name) for c in
+            self.__compositekey__]
+            )
+        self['hash_key'] = raw_sql("UNHEX(SHA1(CONCAT_WS(',', %s)))" % values)
+
     @classproperty
     def trigger_name(cls):
         return 'created_timestamp_' + cls.__tablename__
@@ -173,18 +184,12 @@ class Table(object):
         """ Create this table. Override this method to also create
         dependent tables and any related views that do not already exist.
         """
-        try:
-            # If this uses the staging table or similar, we can
-            # automatically build this here too.
-            cls.__source__.build()
-        except AttributeError:
-            pass
         cls.create_table()
         cls.create_trigger()
 
     @classmethod
     def create_table(cls):
-        """ Create this table in the current data warehouse.
+        """ Create this table in the current datawarehouse.
 
         Returns:
             True if the table was created, or False if the table already
